+++ crates/lexgen/benches/benchmarks.rs
@@ -1,10 +1,9 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 
 fn compile_bench(c: &mut Criterion) {
-    let bench_data = std::fs::read_to_string("bench_data").unwrap();
-    let lexer_ast = syn::parse_str::<lexgen::ast::Lexer>(&bench_data).unwrap();
+    let bench_data = std::fs::read_to_string("/home/omer/rust/lexgen/bench_data").unwrap();
     c.bench_function("compile", |b| {
-        b.iter(|| lexgen::lexer(black_box(lexer_ast.clone())))
+        b.iter(|| lexgen::lexer(black_box(&bench_data)))
     });
 }
 
@@ -13,4 +12,5 @@ criterion_group! {
     config = Criterion::default().sample_size(10);
     targets = compile_bench
 }
+
 criterion_main!(benches);