+++ crates/lexgen/tests/bugs.rs
@@ -22,7 +22,7 @@ fn failure_confusion_1() {
 
         '"' => |lexer| {
             println!("matched a double quote");
-            let str = std::mem::replace(&mut lexer.state().buf, String::new());
+            let str = std::mem::take(&mut lexer.state().buf);
             lexer.return_(str)
         },
 
@@ -75,7 +75,7 @@ fn failure_confusion_2() {
         rule Comment {
             "(*" => |lexer| {
                 let depth = &mut lexer.state().comment_depth;
-                *depth =  *depth + 1;
+                *depth += 1;
                 lexer.continue_()
             },
 
@@ -84,7 +84,7 @@ fn failure_confusion_2() {
                 if *depth == 1 {
                     lexer.switch(LexerRule::Init)
                 } else {
-                    *depth = *depth - 1;
+                    *depth -= 1;
                     lexer.continue_()
                 }
             },
@@ -401,4 +401,38 @@ fn failure_should_reset_state_issue_48() {
     assert_eq!(lexer.next(), Some(Ok((loc(0, 3, 3), "s", loc(0, 4, 4)))));
     assert_eq!(lexer.next(), Some(Ok((loc(0, 4, 4), "a", loc(0, 5, 5)))));
     assert_eq!(lexer.next(), None);
+}
+
+#[test]
+fn new_methods_no_default() {
+    // #54: `new_with_state` and `new_from_iter_with_state` shouldn't require state to implement
+    // `Default`
+
+    struct UserState {}
+
+    lexer! {
+        Lexer(UserState) -> ();
+
+        $ = (),
+    }
+
+    Lexer::new_with_state("", UserState {});
+    Lexer::new_from_iter_with_state(std::iter::empty(), UserState {});
+}
+
+#[test]
+fn new_methods_default() {
+    // #54: `new` and `new_from_iter` should work with user state that implements `Default`
+
+    #[derive(Default)]
+    struct UserState {}
+
+    lexer! {
+        Lexer(UserState) -> ();
+
+        $ = (),
+    }
+
+    Lexer::new("");
+    Lexer::new_from_iter(std::iter::empty());
 }